feat: add new types for entities
