+++ src/shared/types/categoriesTypes.ts
@@ -1,29 +1,7 @@
-export type CategoryType = 'characters' | 'locations' | 'episodes'
+import { Character } from './Character'
+import { Episode } from './Episode'
+import { Location } from './Location'
+
+export type CategoryType = 'character' | 'location' | 'episode'
 
 export type CategoryItem = Character | Episode | Location
-export interface Character {
-  id: number
-  name: string
-  status: string
-  species: string
-  type: string
-  gender: string
-  image: string
-  created: string
-}
-
-export interface Episode {
-  id: number
-  name: string
-  air_date: string
-  episode: string
-  created: string
-}
-
-export interface Location {
-  id: number
-  name: string
-  type: string
-  dimension: string
-  created: string
-}
